Replace aarch64 WFE instruction usage with ISB.
WFE stalls the process for an indeterminate period of time, yielding
slower perfomance when more threads are active than the actual core count.

As a side effect of implementation, ISB stalls execution for some tens
of cycles which (when executed in a loop) decreases memory contention and
thus improves multithreaded performance.

Resolves #48.